requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
streamlit-mermaid>=0.0.3
Pillow>=10.0.0
python-dotenv>=1.0.0